move action types into separate file
