+++ app/scripts/reducers/providers.js
@@ -47,7 +47,7 @@ import {
   OPTIONS_PROVIDERGROUP,
   OPTIONS_PROVIDERGROUP_INFLIGHT,
   OPTIONS_PROVIDERGROUP_ERROR
-} from '../actions';
+} from '../actions/types';
 
 export const initialState = {
   list: {
