+++ app/scripts/reducers/reconciliation-reports.js
@@ -16,7 +16,7 @@ import {
 
   NEW_RECONCILIATION_INFLIGHT,
   NEW_RECONCILIATION
-} from '../actions';
+} from '../actions/types';
 
 export const initialState = {
   list: {
