+++ app/scripts/reducers/rules.js
@@ -36,7 +36,7 @@ import {
   RULE_DISABLE,
   RULE_DISABLE_INFLIGHT,
   RULE_DISABLE_ERROR
-} from '../actions';
+} from '../actions/types';
 
 export const initialState = {
   list: {
