+++ cypress/integration/main_page_spec.js
@@ -4,7 +4,7 @@ import {
   shouldHaveDeletedToken
 } from '../support/assertions';
 
-import { API_VERSION } from '../../app/scripts/actions';
+import { API_VERSION } from '../../app/scripts/actions/types';
 
 describe('Dashboard Home Page', () => {
   it('When not logged in it should redirect to login page', () => {
